feat: add basic signup
