+++ controller/userController.js
@@ -1,0 +1,31 @@
+const bcrypt = require('bcrypt-nodejs')
+const db = require('../models')
+const User = db.User
+
+const userController = {
+  signUpPage: (req, res) => {
+    return res.render('signup')
+  },
+  signUp: (req, res) => {
+    User.create({
+      name: req.body.name,
+      email: req.body.email,
+      password: bcrypt.hashSync(req.body.password, bcrypt.genSaltSync(10), null)
+    }).then(user => {
+      return res.redirect('/signin')
+    })
+  },
+  signInPage: (req, res) => {
+    res.render('signin')
+  },
+  signIn: (req, res) => {
+    req.flash('success_messages', '成功登入！')
+    res.redirect('/restaurants')
+  },
+  logout: (req, res) => {
+    req.flash('success_messages', '登出成功！')
+    req.logout()
+    res.redirect('/signin')
+  }
+}
+module.exports = userController
