+++ views/signup.hbs
@@ -1,0 +1,27 @@
+<form class="form-signin" action="/signup" method="POST" style="width: 100%;">
+  <div class="text-center mb-4">
+    <h1 class="h3 mb-3 font-weight-normal">Sign Up</h1>
+  </div>
+  <div class="form-label-group">
+    <label for="inputName">Name</label>
+    <input type="text" name="name" class="form-control" placeholder="Name" required autofocus />
+  </div>
+  <div class="form-label-group">
+    <label for="inputEmail">Email</label>
+    <input type="text" name="email" class="form-control" placeholder="Email" required autofocus />
+  </div>
+  <div class="form-label-group">
+    <label for="inputPassword">Password</label>
+    <input type="password" name="password" class="form-control" placeholder="Password" required />
+  </div>
+  <div class="form-label-group">
+    <label for="inputPasswordCheck">Password Check</label>
+    <input type="password" name="passwordCheck" class="form-control" placeholder="Password Check" required />
+  </div>
+  <br />
+  <button class="btn btn-lg btn-primary btn-block" type="submit">Submit</button>
+  <div class="text-center mb-4">
+    <p><a href="/signin">Sign In</a></p>
+  </div>
+  <p class="mt-5 mb-3 text-muted text-center">© 2017-2018</p>
+</form>
